requirements.txt
Pillow==8.4.0
pyfiglet==0.8.post1
PySimpleGUI==4.60.4
tqdm==4.65.0
